+++ melpa-recipe.el
@@ -1,0 +1,6 @@
+;; MELPA recipe — submit this to https://github.com/melpa/melpa/tree/master/recipes
+;; File name: sysml2-mode
+(sysml2-mode
+ :fetcher github
+ :repo "jackhale98/sysml2-mode"
+ :files ("*.el" "snippets"))